+++ logic/src/main/java/org/itsallcode/whiterabbit/logic/service/contract/ContractTermsService.java
@@ -1,21 +1,27 @@
 package org.itsallcode.whiterabbit.logic.service.contract;
 
 import java.time.Duration;
+import java.util.Optional;
 
 import org.itsallcode.whiterabbit.logic.Config;
 import org.itsallcode.whiterabbit.logic.model.DayRecord;
 
 public class ContractTermsService
 {
+    public static final Duration BASIC_BREAK = Duration.ofMinutes(45);
     private static final Duration CONTRACTED_HOURS_PER_DAY = Duration.ofHours(8);
     private static final Duration MIN_WORKING_TIME_WITHOUT_BREAK = Duration.ofHours(6);
-    public static final Duration BASIC_BREAK = Duration.ofMinutes(45);
 
-    private final Config config;
+    private final Optional<Duration> hoursPerDay;
 
     public ContractTermsService(Config config)
     {
-        this.config = config;
+        this(config.getCurrentHoursPerDay());
+    }
+
+    public ContractTermsService(Optional<Duration> hoursPerDay)
+    {
+        this.hoursPerDay = hoursPerDay;
     }
 
     public Duration getMandatoryBreak(DayRecord day)
@@ -39,7 +45,7 @@ public class ContractTermsService
 
     public Duration getCurrentWorkingTimePerDay()
     {
-        return config.getCurrentHoursPerDay().orElse(getContractedWorkingTimePerDay());
+        return hoursPerDay.orElse(getContractedWorkingTimePerDay());
     }
 
     public Duration getMandatoryWorkingTime(DayRecord day)