+++ logic/src/main/java/org/itsallcode/whiterabbit/logic/service/project/ProjectService.java
@@ -20,7 +20,6 @@ import javax.json.bind.JsonbBuilder;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
-import org.itsallcode.whiterabbit.logic.Config;
 
 public class ProjectService
 {
@@ -31,16 +30,15 @@ public class ProjectService
     private final Map<String, ProjectImpl> projectsById;
     private final Map<String, ProjectImpl> projectsByLabel;
 
-    public ProjectService(Config config)
+    public ProjectService(Path projectConfigFile)
     {
-        this(JsonbBuilder.create(), config);
+        this(JsonbBuilder.create(), projectConfigFile);
     }
 
-    ProjectService(Jsonb jsonb, Config config)
+    ProjectService(Jsonb jsonb, Path projectConfigFile)
     {
         this.jsonb = jsonb;
 
-        final Path projectConfigFile = config.getProjectFile();
         if (projectConfigFile != null && Files.exists(projectConfigFile))
         {
             final List<ProjectImpl> allProjects = loadAvailableProjects(projectConfigFile);